backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for project management application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "sqlmodel",
    "uvicorn[standard]",
    "httpx",
    "psycopg[binary]",
    "alembic",
    "python-dotenv",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "pytest",
    "pytest-asyncio",
    "coverage",
    "pydantic_settings"

]
